third_party/WebKit/LayoutTests/imported/wpt/html/dom/reflection-misc-expected.txt

This is a testharness.js-based test.
PASS html.title: 77 tests
PASS html.lang: 77 tests
PASS html.dir: 151 tests
PASS html.className (<html class>): 77 tests
PASS html.hidden: 79 tests
PASS html.accessKey: 77 tests
PASS html.tabIndex: 10 tests
FAIL html.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS html.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL html.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS html.tabIndex: 36 tests
FAIL html.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS html.tabIndex: 2 tests
FAIL html.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS html.version: 77 tests
PASS script.title: 77 tests
PASS script.lang: 77 tests
PASS script.dir: 151 tests
PASS script.className (<script class>): 77 tests
PASS script.hidden: 79 tests
PASS script.accessKey: 77 tests
PASS script.tabIndex: 10 tests
FAIL script.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS script.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL script.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS script.tabIndex: 36 tests
FAIL script.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS script.tabIndex: 2 tests
FAIL script.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS script.src: 92 tests
PASS script.type: 77 tests
PASS script.charset: 77 tests
PASS script.defer: 79 tests
PASS script.crossOrigin: 127 tests
PASS noscript.title: 77 tests
PASS noscript.lang: 77 tests
PASS noscript.dir: 151 tests
PASS noscript.className (<noscript class>): 77 tests
PASS noscript.hidden: 79 tests
PASS noscript.accessKey: 77 tests
PASS noscript.tabIndex: 10 tests
FAIL noscript.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS noscript.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL noscript.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS noscript.tabIndex: 36 tests
FAIL noscript.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS noscript.tabIndex: 2 tests
FAIL noscript.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS ins.title: 77 tests
PASS ins.lang: 77 tests
PASS ins.dir: 151 tests
PASS ins.className (<ins class>): 77 tests
PASS ins.hidden: 79 tests
PASS ins.accessKey: 77 tests
PASS ins.tabIndex: 10 tests
FAIL ins.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS ins.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL ins.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS ins.tabIndex: 36 tests
FAIL ins.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS ins.tabIndex: 2 tests
FAIL ins.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS ins.cite: 92 tests
PASS ins.dateTime: 77 tests
PASS del.title: 77 tests
PASS del.lang: 77 tests
PASS del.dir: 151 tests
PASS del.className (<del class>): 77 tests
PASS del.hidden: 79 tests
PASS del.accessKey: 77 tests
PASS del.tabIndex: 10 tests
FAIL del.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS del.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL del.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS del.tabIndex: 36 tests
FAIL del.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS del.tabIndex: 2 tests
FAIL del.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS del.cite: 92 tests
PASS del.dateTime: 77 tests
PASS details.title: 77 tests
PASS details.lang: 77 tests
PASS details.dir: 151 tests
PASS details.className (<details class>): 77 tests
PASS details.hidden: 79 tests
PASS details.accessKey: 77 tests
PASS details.tabIndex: 10 tests
FAIL details.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS details.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL details.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS details.tabIndex: 36 tests
FAIL details.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS details.tabIndex: 2 tests
FAIL details.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS details.open: 79 tests
PASS summary.title: 77 tests
PASS summary.lang: 77 tests
PASS summary.dir: 151 tests
PASS summary.className (<summary class>): 77 tests
PASS summary.hidden: 79 tests
PASS summary.accessKey: 77 tests
PASS summary.tabIndex: 2 tests
FAIL summary.tabIndex: setAttribute() to -36 followed by IDL get assert_equals: expected -36 but got -1
PASS summary.tabIndex: 3 tests
FAIL summary.tabIndex: setAttribute() to 0 followed by IDL get assert_equals: expected 0 but got -1
PASS summary.tabIndex: setAttribute() to 1 followed by getAttribute() 
FAIL summary.tabIndex: setAttribute() to 1 followed by IDL get assert_equals: expected 1 but got -1
PASS summary.tabIndex: setAttribute() to 2147483647 followed by getAttribute() 
FAIL summary.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got -1
PASS summary.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL summary.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -1
PASS summary.tabIndex: 3 tests
FAIL summary.tabIndex: setAttribute() to "-0" followed by IDL get assert_equals: expected 0 but got -1
PASS summary.tabIndex: setAttribute() to "0" followed by getAttribute() 
FAIL summary.tabIndex: setAttribute() to "0" followed by IDL get assert_equals: expected 0 but got -1
PASS summary.tabIndex: setAttribute() to "1" followed by getAttribute() 
FAIL summary.tabIndex: setAttribute() to "1" followed by IDL get assert_equals: expected 1 but got -1
PASS summary.tabIndex: setAttribute() to "\t7" followed by getAttribute() 
FAIL summary.tabIndex: setAttribute() to "\t7" followed by IDL get assert_equals: expected 7 but got -1
PASS summary.tabIndex: setAttribute() to "\f7" followed by getAttribute() 
FAIL summary.tabIndex: setAttribute() to "\f7" followed by IDL get assert_equals: expected 7 but got -1
PASS summary.tabIndex: setAttribute() to " 7" followed by getAttribute() 
FAIL summary.tabIndex: setAttribute() to " 7" followed by IDL get assert_equals: expected 7 but got -1
PASS summary.tabIndex: setAttribute() to "\n7" followed by getAttribute() 
FAIL summary.tabIndex: setAttribute() to "\n7" followed by IDL get assert_equals: expected 7 but got -1
PASS summary.tabIndex: setAttribute() to "\r7" followed by getAttribute() 
FAIL summary.tabIndex: setAttribute() to "\r7" followed by IDL get assert_equals: expected 7 but got -1
PASS summary.tabIndex: setAttribute() to 1.5 followed by getAttribute() 
FAIL summary.tabIndex: setAttribute() to 1.5 followed by IDL get assert_equals: expected 1 but got -1
PASS summary.tabIndex: setAttribute() to object "2" followed by getAttribute() 
FAIL summary.tabIndex: setAttribute() to object "2" followed by IDL get assert_equals: expected 2 but got -1
PASS summary.tabIndex: 2 tests
FAIL summary.tabIndex: IDL set to -36 followed by IDL get assert_equals: expected -36 but got -1
PASS summary.tabIndex: 5 tests
FAIL summary.tabIndex: IDL set to 0 followed by IDL get assert_equals: expected 0 but got -1
PASS summary.tabIndex: 2 tests
FAIL summary.tabIndex: IDL set to 1 followed by IDL get assert_equals: expected 1 but got -1
PASS summary.tabIndex: 2 tests
FAIL summary.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got -1
PASS summary.tabIndex: 2 tests
FAIL summary.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -1
PASS menu.title: 77 tests
PASS menu.lang: 77 tests
PASS menu.dir: 151 tests
PASS menu.className (<menu class>): 77 tests
PASS menu.hidden: 79 tests
PASS menu.accessKey: 77 tests
PASS menu.tabIndex: 10 tests
FAIL menu.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS menu.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL menu.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS menu.tabIndex: 36 tests
FAIL menu.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS menu.tabIndex: 2 tests
FAIL menu.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS menu.type: typeof IDL attribute 
FAIL menu.type: IDL get with DOM attribute unset assert_equals: expected "toolbar" but got ""
PASS menu.type: setAttribute() to "" followed by getAttribute() 
FAIL menu.type: setAttribute() to "" followed by IDL get assert_equals: expected "toolbar" but got ""
PASS menu.type: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " followed by getAttribute() 
FAIL menu.type: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " followed by IDL get assert_equals: expected "toolbar" but got " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo "
PASS menu.type: setAttribute() to undefined followed by getAttribute() 
FAIL menu.type: setAttribute() to undefined followed by IDL get assert_equals: expected "toolbar" but got "undefined"
PASS menu.type: setAttribute() to 7 followed by getAttribute() 
FAIL menu.type: setAttribute() to 7 followed by IDL get assert_equals: expected "toolbar" but got "7"
PASS menu.type: setAttribute() to 1.5 followed by getAttribute() 
FAIL menu.type: setAttribute() to 1.5 followed by IDL get assert_equals: expected "toolbar" but got "1.5"
PASS menu.type: setAttribute() to true followed by getAttribute() 
FAIL menu.type: setAttribute() to true followed by IDL get assert_equals: expected "toolbar" but got "true"
PASS menu.type: setAttribute() to false followed by getAttribute() 
FAIL menu.type: setAttribute() to false followed by IDL get assert_equals: expected "toolbar" but got "false"
PASS menu.type: setAttribute() to object "[object Object]" followed by getAttribute() 
FAIL menu.type: setAttribute() to object "[object Object]" followed by IDL get assert_equals: expected "toolbar" but got "[object Object]"
PASS menu.type: setAttribute() to NaN followed by getAttribute() 
FAIL menu.type: setAttribute() to NaN followed by IDL get assert_equals: expected "toolbar" but got "NaN"
PASS menu.type: setAttribute() to Infinity followed by getAttribute() 
FAIL menu.type: setAttribute() to Infinity followed by IDL get assert_equals: expected "toolbar" but got "Infinity"
PASS menu.type: setAttribute() to -Infinity followed by getAttribute() 
FAIL menu.type: setAttribute() to -Infinity followed by IDL get assert_equals: expected "toolbar" but got "-Infinity"
PASS menu.type: setAttribute() to "\0" followed by getAttribute() 
FAIL menu.type: setAttribute() to "\0" followed by IDL get assert_equals: expected "toolbar" but got "\0"
PASS menu.type: setAttribute() to null followed by getAttribute() 
FAIL menu.type: setAttribute() to null followed by IDL get assert_equals: expected "toolbar" but got "null"
PASS menu.type: setAttribute() to object "test-toString" followed by getAttribute() 
FAIL menu.type: setAttribute() to object "test-toString" followed by IDL get assert_equals: expected "toolbar" but got "test-toString"
PASS menu.type: setAttribute() to object "test-valueOf" followed by getAttribute() 
FAIL menu.type: setAttribute() to object "test-valueOf" followed by IDL get assert_equals: expected "toolbar" but got "test-valueOf"
PASS menu.type: 3 tests
FAIL menu.type: setAttribute() to "xpopup" followed by IDL get assert_equals: expected "toolbar" but got "xpopup"
PASS menu.type: setAttribute() to "popup\0" followed by getAttribute() 
FAIL menu.type: setAttribute() to "popup\0" followed by IDL get assert_equals: expected "toolbar" but got "popup\0"
PASS menu.type: setAttribute() to "opup" followed by getAttribute() 
FAIL menu.type: setAttribute() to "opup" followed by IDL get assert_equals: expected "toolbar" but got "opup"
PASS menu.type: setAttribute() to "POPUP" followed by getAttribute() 
FAIL menu.type: setAttribute() to "POPUP" followed by IDL get assert_equals: expected "popup" but got "POPUP"
PASS menu.type: 3 tests
FAIL menu.type: setAttribute() to "xtoolbar" followed by IDL get assert_equals: expected "toolbar" but got "xtoolbar"
PASS menu.type: setAttribute() to "toolbar\0" followed by getAttribute() 
FAIL menu.type: setAttribute() to "toolbar\0" followed by IDL get assert_equals: expected "toolbar" but got "toolbar\0"
PASS menu.type: setAttribute() to "oolbar" followed by getAttribute() 
FAIL menu.type: setAttribute() to "oolbar" followed by IDL get assert_equals: expected "toolbar" but got "oolbar"
PASS menu.type: setAttribute() to "TOOLBAR" followed by getAttribute() 
FAIL menu.type: setAttribute() to "TOOLBAR" followed by IDL get assert_equals: expected "toolbar" but got "TOOLBAR"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to "" followed by IDL get assert_equals: expected "toolbar" but got ""
PASS menu.type: 2 tests
FAIL menu.type: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " followed by IDL get assert_equals: expected "toolbar" but got " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo "
PASS menu.type: 2 tests
FAIL menu.type: IDL set to undefined followed by IDL get assert_equals: expected "toolbar" but got "undefined"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to 7 followed by IDL get assert_equals: expected "toolbar" but got "7"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to 1.5 followed by IDL get assert_equals: expected "toolbar" but got "1.5"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to true followed by IDL get assert_equals: expected "toolbar" but got "true"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to false followed by IDL get assert_equals: expected "toolbar" but got "false"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to object "[object Object]" followed by IDL get assert_equals: expected "toolbar" but got "[object Object]"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to NaN followed by IDL get assert_equals: expected "toolbar" but got "NaN"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to Infinity followed by IDL get assert_equals: expected "toolbar" but got "Infinity"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to -Infinity followed by IDL get assert_equals: expected "toolbar" but got "-Infinity"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to "\0" followed by IDL get assert_equals: expected "toolbar" but got "\0"
PASS menu.type: IDL set to null should not throw 
FAIL menu.type: IDL set to null followed by IDL get assert_equals: expected "toolbar" but got "null"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to object "test-toString" followed by IDL get assert_equals: expected "toolbar" but got "test-toString"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to object "test-valueOf" followed by IDL get assert_equals: expected "toolbar" but got "test-valueOf"
PASS menu.type: 5 tests
FAIL menu.type: IDL set to "xpopup" followed by IDL get assert_equals: expected "toolbar" but got "xpopup"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to "popup\0" followed by IDL get assert_equals: expected "toolbar" but got "popup\0"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to "opup" followed by IDL get assert_equals: expected "toolbar" but got "opup"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to "POPUP" followed by IDL get assert_equals: expected "popup" but got "POPUP"
PASS menu.type: 5 tests
FAIL menu.type: IDL set to "xtoolbar" followed by IDL get assert_equals: expected "toolbar" but got "xtoolbar"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to "toolbar\0" followed by IDL get assert_equals: expected "toolbar" but got "toolbar\0"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to "oolbar" followed by IDL get assert_equals: expected "toolbar" but got "oolbar"
PASS menu.type: 2 tests
FAIL menu.type: IDL set to "TOOLBAR" followed by IDL get assert_equals: expected "toolbar" but got "TOOLBAR"
PASS menu.label: 77 tests
PASS menu.compact: 79 tests
PASS menuitem.title: 77 tests
PASS menuitem.lang: 77 tests
PASS menuitem.dir: 151 tests
PASS menuitem.className (<menuitem class>): 77 tests
PASS menuitem.hidden: 79 tests
PASS menuitem.accessKey: 77 tests
PASS menuitem.tabIndex: 10 tests
FAIL menuitem.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS menuitem.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL menuitem.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS menuitem.tabIndex: 36 tests
FAIL menuitem.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS menuitem.tabIndex: 2 tests
FAIL menuitem.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS menuitem.type: typeof IDL attribute 
FAIL menuitem.type: IDL get with DOM attribute unset assert_equals: expected "command" but got ""
PASS menuitem.type: setAttribute() to "" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to "" followed by IDL get assert_equals: expected "command" but got ""
PASS menuitem.type: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " followed by getAttribute() 
FAIL menuitem.type: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " followed by IDL get assert_equals: expected "command" but got " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo "
PASS menuitem.type: setAttribute() to undefined followed by getAttribute() 
FAIL menuitem.type: setAttribute() to undefined followed by IDL get assert_equals: expected "command" but got "undefined"
PASS menuitem.type: setAttribute() to 7 followed by getAttribute() 
FAIL menuitem.type: setAttribute() to 7 followed by IDL get assert_equals: expected "command" but got "7"
PASS menuitem.type: setAttribute() to 1.5 followed by getAttribute() 
FAIL menuitem.type: setAttribute() to 1.5 followed by IDL get assert_equals: expected "command" but got "1.5"
PASS menuitem.type: setAttribute() to true followed by getAttribute() 
FAIL menuitem.type: setAttribute() to true followed by IDL get assert_equals: expected "command" but got "true"
PASS menuitem.type: setAttribute() to false followed by getAttribute() 
FAIL menuitem.type: setAttribute() to false followed by IDL get assert_equals: expected "command" but got "false"
PASS menuitem.type: setAttribute() to object "[object Object]" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to object "[object Object]" followed by IDL get assert_equals: expected "command" but got "[object Object]"
PASS menuitem.type: setAttribute() to NaN followed by getAttribute() 
FAIL menuitem.type: setAttribute() to NaN followed by IDL get assert_equals: expected "command" but got "NaN"
PASS menuitem.type: setAttribute() to Infinity followed by getAttribute() 
FAIL menuitem.type: setAttribute() to Infinity followed by IDL get assert_equals: expected "command" but got "Infinity"
PASS menuitem.type: setAttribute() to -Infinity followed by getAttribute() 
FAIL menuitem.type: setAttribute() to -Infinity followed by IDL get assert_equals: expected "command" but got "-Infinity"
PASS menuitem.type: setAttribute() to "\0" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to "\0" followed by IDL get assert_equals: expected "command" but got "\0"
PASS menuitem.type: setAttribute() to null followed by getAttribute() 
FAIL menuitem.type: setAttribute() to null followed by IDL get assert_equals: expected "command" but got "null"
PASS menuitem.type: setAttribute() to object "test-toString" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to object "test-toString" followed by IDL get assert_equals: expected "command" but got "test-toString"
PASS menuitem.type: setAttribute() to object "test-valueOf" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to object "test-valueOf" followed by IDL get assert_equals: expected "command" but got "test-valueOf"
PASS menuitem.type: 3 tests
FAIL menuitem.type: setAttribute() to "xcommand" followed by IDL get assert_equals: expected "command" but got "xcommand"
PASS menuitem.type: setAttribute() to "command\0" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to "command\0" followed by IDL get assert_equals: expected "command" but got "command\0"
PASS menuitem.type: setAttribute() to "ommand" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to "ommand" followed by IDL get assert_equals: expected "command" but got "ommand"
PASS menuitem.type: setAttribute() to "COMMAND" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to "COMMAND" followed by IDL get assert_equals: expected "command" but got "COMMAND"
PASS menuitem.type: 3 tests
FAIL menuitem.type: setAttribute() to "xcheckbox" followed by IDL get assert_equals: expected "command" but got "xcheckbox"
PASS menuitem.type: setAttribute() to "checkbox\0" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to "checkbox\0" followed by IDL get assert_equals: expected "command" but got "checkbox\0"
PASS menuitem.type: setAttribute() to "heckbox" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to "heckbox" followed by IDL get assert_equals: expected "command" but got "heckbox"
PASS menuitem.type: setAttribute() to "CHECKBOX" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to "CHECKBOX" followed by IDL get assert_equals: expected "checkbox" but got "CHECKBOX"
PASS menuitem.type: 3 tests
FAIL menuitem.type: setAttribute() to "xradio" followed by IDL get assert_equals: expected "command" but got "xradio"
PASS menuitem.type: setAttribute() to "radio\0" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to "radio\0" followed by IDL get assert_equals: expected "command" but got "radio\0"
PASS menuitem.type: setAttribute() to "adio" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to "adio" followed by IDL get assert_equals: expected "command" but got "adio"
PASS menuitem.type: setAttribute() to "RADIO" followed by getAttribute() 
FAIL menuitem.type: setAttribute() to "RADIO" followed by IDL get assert_equals: expected "radio" but got "RADIO"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to "" followed by IDL get assert_equals: expected "command" but got ""
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " followed by IDL get assert_equals: expected "command" but got " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo "
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to undefined followed by IDL get assert_equals: expected "command" but got "undefined"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to 7 followed by IDL get assert_equals: expected "command" but got "7"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to 1.5 followed by IDL get assert_equals: expected "command" but got "1.5"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to true followed by IDL get assert_equals: expected "command" but got "true"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to false followed by IDL get assert_equals: expected "command" but got "false"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to object "[object Object]" followed by IDL get assert_equals: expected "command" but got "[object Object]"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to NaN followed by IDL get assert_equals: expected "command" but got "NaN"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to Infinity followed by IDL get assert_equals: expected "command" but got "Infinity"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to -Infinity followed by IDL get assert_equals: expected "command" but got "-Infinity"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to "\0" followed by IDL get assert_equals: expected "command" but got "\0"
PASS menuitem.type: IDL set to null should not throw 
FAIL menuitem.type: IDL set to null followed by IDL get assert_equals: expected "command" but got "null"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to object "test-toString" followed by IDL get assert_equals: expected "command" but got "test-toString"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to object "test-valueOf" followed by IDL get assert_equals: expected "command" but got "test-valueOf"
PASS menuitem.type: 5 tests
FAIL menuitem.type: IDL set to "xcommand" followed by IDL get assert_equals: expected "command" but got "xcommand"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to "command\0" followed by IDL get assert_equals: expected "command" but got "command\0"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to "ommand" followed by IDL get assert_equals: expected "command" but got "ommand"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to "COMMAND" followed by IDL get assert_equals: expected "command" but got "COMMAND"
PASS menuitem.type: 5 tests
FAIL menuitem.type: IDL set to "xcheckbox" followed by IDL get assert_equals: expected "command" but got "xcheckbox"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to "checkbox\0" followed by IDL get assert_equals: expected "command" but got "checkbox\0"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to "heckbox" followed by IDL get assert_equals: expected "command" but got "heckbox"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to "CHECKBOX" followed by IDL get assert_equals: expected "checkbox" but got "CHECKBOX"
PASS menuitem.type: 5 tests
FAIL menuitem.type: IDL set to "xradio" followed by IDL get assert_equals: expected "command" but got "xradio"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to "radio\0" followed by IDL get assert_equals: expected "command" but got "radio\0"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to "adio" followed by IDL get assert_equals: expected "command" but got "adio"
PASS menuitem.type: 2 tests
FAIL menuitem.type: IDL set to "RADIO" followed by IDL get assert_equals: expected "radio" but got "RADIO"
PASS menuitem.label: 77 tests
PASS menuitem.icon: 92 tests
PASS menuitem.disabled: 79 tests
PASS menuitem.checked: 79 tests
PASS menuitem.radiogroup: 77 tests
PASS menuitem.default: 79 tests
PASS dialog.title: 77 tests
PASS dialog.lang: 77 tests
PASS dialog.dir: 151 tests
PASS dialog.className (<dialog class>): 77 tests
PASS dialog.hidden: 79 tests
PASS dialog.accessKey: 77 tests
PASS dialog.tabIndex: 10 tests
FAIL dialog.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS dialog.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL dialog.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS dialog.tabIndex: 36 tests
FAIL dialog.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS dialog.tabIndex: 2 tests
FAIL dialog.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS dialog.open: 79 tests
PASS undefinedelement.title: 77 tests
PASS undefinedelement.lang: 77 tests
PASS undefinedelement.dir: 151 tests
PASS undefinedelement.className (<undefinedelement class>): 77 tests
PASS undefinedelement.hidden: 79 tests
PASS undefinedelement.accessKey: 77 tests
PASS undefinedelement.tabIndex: 10 tests
FAIL undefinedelement.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS undefinedelement.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL undefinedelement.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS undefinedelement.tabIndex: 36 tests
FAIL undefinedelement.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS undefinedelement.tabIndex: 2 tests
FAIL undefinedelement.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
Harness: the test ran to completion.
